commint

+++ src/com/gogo/map/GoMapHelper.java
@@ -7,10 +7,10 @@ public class GoMapHelper {
 	public static final float COORD_RANGE = 100.0f;
 	
 	//经度
-	private float longitude;
+	private double longitude;
 	
 	//纬度
-	private float latitude;
+	private double latitude;
 	/**
 	 * 根据ip得到城市信息
 	 * @param addr
@@ -20,16 +20,17 @@ public class GoMapHelper {
 		//TODO 通过ip获得城市信息
 		return null;
 	}
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
-	public float getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
+	
 }
